+++ flash_attn/flash_attn_triton_amd/__init__.py
@@ -1,4 +1,4 @@
-from . import interface_fa as flash_attn_2
-from . import interface_fa_v3 as flash_attn_3
+from . import interface_v2 as flash_attn_2
+from . import interface_v3 as flash_attn_3
 
 __all__ = ["flash_attn_2", "flash_attn_3"]
